# openapi/backend-api.yaml
openapi: 3.1.0
info:
  title: Refex User Engagement Report API
  version: 1.0.0
  description: |
    Authoritative backend API for the Admin UI.
    Frontend must consume this contract only — no direct Cloud SQL, BigQuery, Kissflow, or SMTP access.
servers:
  - url: /api/v1
    description: Backend API (Cloud Run)
...
paths:
...
  /applications:
    get:
      tags: [applications]
      summary: List connected applications
      responses:
        '200':
          description: Application list
    post:
      tags: [applications]
      summary: Register application (credential ref only)
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationRegistrationRequest'
      responses:
        '201':
          description: Created
        '200':
          description: Idempotent replay
        '409':
          description: Conflict
  /applications/{applicationId}/processes:
    get:
      tags: [applications]
      summary: List processes/resources for application
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Process list
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: true
      schema:
        type: string
        minLength: 8
  schemas:
...
    ApplicationRegistrationRequest:
      type: object
      required:
        - kissflow_account_id
        - application_id
        - subdomain
        - access_key_id
        - access_key_secret
        - environment
      properties:
        kissflow_account_id:
          type: string
        application_id:
          type: string
        application_name:
          type: string
        display_name:
          type: string
        subdomain:
          type: string
        region:
          type: string
          default: com
        environment:
          type: string
          example: production
        description:
          type: string
        access_key_id:
          type: string
          description: Accepted in request body; stored as Secret Manager ref only
        access_key_secret:
          type: string
          description: Accepted in request body; never persisted in PostgreSQL
        process_ids:
          type: array
          items:
            type: string
        dataform_ids:
          type: array
          items:
            type: string
        board_ids:
          type: array
          items:
            type: string
        dataset_ids:
          type: array
          items:
            type: string